--- app/src/main/java/com/practicum/playlistmakerapp/data/network/ITunesApi.kt
@@ -1,10 +1,0 @@
-package com.practicum.playlistmakerapp.data.network
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ITunesApi {
-    @GET("/search?entity=song")
-    fun trackSearch(@Query("term") text: String): Call<TrackResponse>
-}
