--- app/src/main/java/com/practicum/playlistmakerapp/data/network/TrackResponse.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmakerapp.data.network
-
-import com.practicum.playlistmakerapp.domain.models.Track
-
-data class TrackResponse(
-    val resultCount: String,
-    val results: List<Track>)
