--- app/src/main/java/com/practicum/playlistmakerapp/domain/models/Track.kt
@@ -1,19 +1,0 @@
-package com.practicum.playlistmakerapp.domain.models
-
-data class Track(
-    val trackId: String,
-    val trackName: String,
-    val artistName: String,
-    val trackTimeMillis: String,
-    val artworkUrl100: String,
-    val collectionName: String,
-    val releaseDate: String,
-    val primaryGenreName: String,
-    val country: String,
-    val previewUrl: String,
-    var isPlaying: Boolean = false
-) {
-    fun getCoverArtwork(): String {
-        return artworkUrl100.replaceAfterLast('/', "512x512bb.jpg")
-    }
-}
