--- app/src/main/java/com/practicum/playlistmakerapp/domain/repository/SearchHistoryRepository.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmakerapp.domain.repository
-
-import com.practicum.playlistmakerapp.domain.models.Track
-
-interface SearchHistoryRepository {
-    fun addTrack(track: Track)
-    fun clearSearchHistory()
-    fun getTracks(): ArrayList<Track>
-}
