--- app/src/main/java/com/practicum/playlistmakerapp/domain/usecases/PauseTrackUseCase.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmakerapp.domain.usecases
-
-import com.practicum.playlistmakerapp.domain.repository.TrackRepository
-
-class PauseTrackUseCase(private val trackRepository: TrackRepository) {
-    fun pause() {
-        trackRepository.pauseTrack()
-    }
-}
