--- app/src/main/java/com/practicum/playlistmakerapp/domain/usecases/PlayTrackUseCase.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmakerapp.domain.usecases
-
-import com.practicum.playlistmakerapp.domain.repository.TrackRepository
-
-class PlayTrackUseCase(private val trackRepository: TrackRepository) {
-    fun play() {
-        trackRepository.playTrack()
-    }
-}
