--- app/src/main/java/com/practicum/playlistmakerapp/domain/usecases/UpdateTrackPositionUseCase.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmakerapp.domain.usecases
-
-import com.practicum.playlistmakerapp.domain.repository.TrackRepository
-
-class UpdateTrackPositionUseCase(private val trackRepository: TrackRepository) : UseCase<Unit, Int> {
-    override fun execute(params: Unit): Int {
-        return trackRepository.getTrackPosition()
-    }
-}
