--- app/src/main/java/com/practicum/playlistmakerapp/domain/usecases/UseCase.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmakerapp.domain.usecases
-
-interface UseCase<in Params, out Type> {
-    fun execute(params: Params): Type
-}
