+++ app/src/main/java/com/practicum/playlistmakerapp/main/ui/MainActivity.kt
@@ -1,0 +1,37 @@
+package com.practicum.playlistmakerapp.main.ui
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.Button
+import com.practicum.playlistmakerapp.R
+import com.practicum.playlistmakerapp.media.ui.MediaActivity
+import com.practicum.playlistmakerapp.search.ui.SearchActivity
+import com.practicum.playlistmakerapp.settings.ui.SettingsActivity
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        val searchButton = findViewById<Button>(R.id.search_button)
+        val mediaButton = findViewById<Button>(R.id.media_button)
+        val settingsButton = findViewById<Button>(R.id.setting_button)
+
+        searchButton.setOnClickListener {
+            val displayIntent = Intent(this, SearchActivity::class.java)
+            startActivity(displayIntent)
+        }
+
+        mediaButton.setOnClickListener {
+            val displayIntent = Intent(this, MediaActivity::class.java)
+            startActivity(displayIntent)
+        }
+
+        settingsButton.setOnClickListener {
+            val displayIntent = Intent(this, SettingsActivity::class.java)
+            startActivity(displayIntent)
+        }
+    }
+}
