+++ app/src/main/java/com/practicum/playlistmakerapp/media/ui/MediaActivity.kt
@@ -1,0 +1,12 @@
+package com.practicum.playlistmakerapp.media.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.practicum.playlistmakerapp.R
+
+class MediaActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_media)
+    }
+}
