+++ app/src/main/java/com/practicum/playlistmakerapp/player/usecases/GetTrackUseCase.kt
@@ -1,0 +1,10 @@
+package com.practicum.playlistmakerapp.player.usecases
+
+import com.practicum.playlistmakerapp.player.domain.repository.TrackRepository
+import com.practicum.playlistmakerapp.player.domain.models.Track
+
+class GetTrackUseCase(private val trackRepository: TrackRepository) : UseCase<String, Track?> {
+    override fun execute(params: String): Track? {
+        return trackRepository.getTrack(params)
+    }
+}
