+++ app/src/main/java/com/practicum/playlistmakerapp/player/usecases/PauseTrackUseCase.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmakerapp.player.usecases
+
+import com.practicum.playlistmakerapp.player.domain.repository.TrackRepository
+
+class PauseTrackUseCase(private val trackRepository: TrackRepository) {
+    fun pause() {
+        trackRepository.pauseTrack()
+    }
+}
