+++ app/src/main/java/com/practicum/playlistmakerapp/player/usecases/PrepareTrackUseCase.kt
@@ -1,0 +1,16 @@
+package com.practicum.playlistmakerapp.player.usecases
+
+import com.practicum.playlistmakerapp.player.domain.repository.TrackRepository
+
+class PrepareTrackUseCase(private val trackRepository: TrackRepository) :
+    UseCase<PrepareTrackUseCase.Params, Unit> {
+    data class Params(
+        val url: String,
+        val onPrepared: () -> Unit,
+        val onError: () -> Unit
+    )
+
+    override fun execute(params: Params) {
+        trackRepository.prepareTrack(params.url, params.onPrepared, params.onError)
+    }
+}
