+++ app/src/main/java/com/practicum/playlistmakerapp/player/usecases/UpdateTrackPositionUseCase.kt
@@ -1,0 +1,10 @@
+package com.practicum.playlistmakerapp.player.usecases
+
+import com.practicum.playlistmakerapp.player.domain.repository.TrackRepository
+
+class UpdateTrackPositionUseCase(private val trackRepository: TrackRepository) :
+    UseCase<Unit, Int> {
+    override fun execute(params: Unit): Int {
+        return trackRepository.getTrackPosition()
+    }
+}
