+++ app/src/main/java/com/practicum/playlistmakerapp/player/usecases/UseCase.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmakerapp.player.usecases
+
+interface UseCase<in Params, out Type> {
+    fun execute(params: Params): Type
+}
