--- app/src/main/java/com/practicum/playlistmakerapp/presentation/ui/media/MediaActivity.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlistmakerapp.presentation.ui.media
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.practicum.playlistmakerapp.R
-
-class MediaActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_media)
-    }
-}
