+++ app/src/main/java/com/practicum/playlistmakerapp/search/domain/repository/SearchHistoryRepository.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmakerapp.search.domain.repository
+
+import com.practicum.playlistmakerapp.player.domain.models.Track
+
+interface SearchHistoryRepository {
+    fun addTrack(track: Track)
+    fun clearSearchHistory()
+    fun getTracks(): ArrayList<Track>
+}
