+++ app/src/main/java/com/practicum/playlistmakerapp/search/ui/TrackViewHolder.kt
@@ -1,0 +1,35 @@
+package com.practicum.playlistmakerapp.search.ui
+
+import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.practicum.playlistmakerapp.R
+import com.practicum.playlistmakerapp.player.domain.models.Track
+import com.practicum.playlistmakerapp.DpToPx
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class TrackViewHolder(itemView: View): RecyclerView.ViewHolder(itemView) {
+    private val trackName: TextView = itemView.findViewById(R.id.trackName)
+    private val trackImage: ImageView = itemView.findViewById(R.id.trackImage)
+    private val artistName: TextView = itemView.findViewById(R.id.artistName)
+    private val trackTimeMillis: TextView = itemView.findViewById(R.id.trackTime)
+
+    fun bind(model: Track) {
+        trackName.text = model.trackName
+
+        Glide.with(trackImage)
+            .load(model.artworkUrl100)
+            .centerCrop()
+            .placeholder(R.drawable.placeholder)
+            .transform(RoundedCorners(DpToPx.dpToPx(2f, trackImage)))
+            .into(trackImage)
+
+        artistName.text = model.artistName
+        trackTimeMillis.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(model.trackTimeMillis.toLong())
+
+    }
+}
