+++ app/src/main/res/layout/activity_media.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.media.MediaActivity">
+    tools:context=".media.ui.MediaActivity">
 
     <TextView
         android:id="@+id/message"
