+++ app/src/main/res/layout/audio_player.xml
@@ -84,8 +84,9 @@
             app:layout_constraintTop_toBottomOf="@+id/trackAuthorTV"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:srcCompat="@drawable/play"
-            app:tint="@color/text" />
+            android:src="@drawable/play"
+            app:tint="@color/text"
+            android:visibility="visible"/>
 
         <!-- Кнопка Пауза -->
         <ImageView
@@ -97,9 +98,9 @@
             app:layout_constraintTop_toBottomOf="@+id/trackAuthorTV"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:srcCompat="@drawable/pause"
+            android:src="@drawable/pause"
             app:tint="@color/text"
-            android:visibility="gone"/>
+            android:visibility="invisible"/>
 
         <!-- Добавить в плейлист -->
         <ImageView
@@ -110,7 +111,7 @@
             app:layout_constraintTop_toBottomOf="@+id/trackAuthorTV"
             app:layout_constraintStart_toStartOf="@id/left_guideline"
             app:layout_constraintEnd_toStartOf="@id/playIV"
-            app:srcCompat="@drawable/plus" />
+            android:src="@drawable/plus" />
 
         <!-- Плейлист создан -->
         <ImageView
@@ -121,7 +122,7 @@
             app:layout_constraintTop_toBottomOf="@+id/trackAuthorTV"
             app:layout_constraintStart_toStartOf="@id/left_guideline"
             app:layout_constraintEnd_toStartOf="@id/playIV"
-            app:srcCompat="@drawable/playlist_created"
+            android:src="@drawable/playlist_created"
             android:visibility="gone"/>
 
         <!-- Добавить в избранное-->
@@ -133,7 +134,7 @@
             app:layout_constraintTop_toBottomOf="@+id/trackAuthorTV"
             app:layout_constraintStart_toEndOf="@id/playIV"
             app:layout_constraintEnd_toEndOf="@id/right_guideline"
-            app:srcCompat="@drawable/add_to_favorites" />
+            android:src="@drawable/add_to_favorites" />
 
         <!-- Добавлен в избранное-->
         <ImageView
@@ -144,7 +145,7 @@
             app:layout_constraintTop_toBottomOf="@+id/trackAuthorTV"
             app:layout_constraintStart_toEndOf="@id/playIV"
             app:layout_constraintEnd_toEndOf="@id/right_guideline"
-            app:srcCompat="@drawable/added_to_favorites"
+            android:src="@drawable/added_to_favorites"
             android:visibility="gone"/>
 
         <!-- Длительность воспроизведения -->
